feat: Built function to take the user's choices and call one of the operator functions, and set the three choice variables as null
The three variables were set to null as they haven't been defined yet by the user. Those values will be provided by the event listeners that I will place yet, which will call  upon the function operate which will call the operaters  and store the result in a variable (result).